Refactoring - create new modules, delete modules, rename modules, and move around functions accordingly.
