+++ harmony_regridding_service/resample.py
@@ -1,4 +1,4 @@
-"""Module for resampling projection data."""
+"""Module for resampling functions."""
 
 from logging import getLogger
 
@@ -12,24 +12,23 @@ from varinfo import VarInfoFromNetCDF4
 
 from harmony_regridding_service.dimensions import (
     _all_dimensions,
+    _copy_1d_dimension_variables,
     _copy_dimensions,
     _create_dimension,
     _get_column_dims,
     _get_row_dims,
     _horizontal_dims_for_variable,
+    _is_column_dim,
+    _is_row_dim,
 )
 from harmony_regridding_service.exceptions import (
     RegridderException,
 )
 from harmony_regridding_service.grid import _compute_source_swath
-from harmony_regridding_service.resample_utilities import (
-    _get_rows_per_scan,
-    _integer_like,
-    _prepare_data_plane,
-)
-from harmony_regridding_service.variable_utilities import (
+from harmony_regridding_service.utilities import (
     _copy_var_with_attrs,
     _copy_var_without_metadata,
+    _integer_like,
 )
 
 logger = getLogger(__name__)
@@ -280,7 +279,7 @@ def _cache_resamplers(
     return grid_cache
 
 
-def _transfer_dimensions(
+def _transfer_resampled_dimensions(
     source_ds: Dataset,
     target_ds: Dataset,
     target_area: AreaDefinition,
@@ -301,3 +300,95 @@ def _transfer_dimensions(
     _create_resampled_dimensions(
         resampled_dimension_pairs, target_ds, target_area, var_info
     )
+
+
+def _copy_resampled_dimension_variables(
+    source_ds: Dataset,
+    target_ds: Dataset,
+    target_area: AreaDefinition,
+    var_info: VarInfoFromNetCDF4,
+) -> set[str]:
+    """Copy over dimension variables that are changed in the target file."""
+    dim_var_names = _resampled_dimension_variable_names(var_info)
+    processed_vars = _copy_1d_dimension_variables(
+        source_ds, target_ds, dim_var_names, target_area, var_info
+    )
+
+    bounds_vars = dim_var_names - processed_vars
+    for bounds_var in bounds_vars:
+        processed_vars |= _copy_resampled_bounds_variable(
+            source_ds, target_ds, bounds_var, target_area, var_info
+        )
+
+    return processed_vars
+
+
+def _get_rows_per_scan(total_rows: int) -> int:
+    """Gets optimum value for rows per scan.
+
+    Finds the smallest divisor of the total number of rows. If no divisor is
+    found, return the total number of rows.
+
+    """
+    if total_rows < 2:
+        return 1
+    for row_number in range(2, int(total_rows**0.5) + 1):
+        if total_rows % row_number == 0:
+            logger.info(f'rows_per_scan = {row_number}')
+            return row_number
+
+    logger.info(f'returning all rows for rows_per_scan = {total_rows}')
+    return total_rows
+
+
+def _prepare_data_plane(
+    data: np.array,
+    var_info: VarInfoFromNetCDF4,
+    var_name: str,
+    cast_to: np.dtype | None,
+) -> np.array:
+    """Perform Type casting and transpose 2d data array when necessary.
+
+    Also perform a transposition if the data dimension organization requires.
+    """
+    if cast_to is not None and data.dtype != cast_to:
+        data = data.astype(cast_to)
+
+    if _needs_rotation(var_info, var_name):
+        data = np.ma.copy(data.T, order='C')
+
+    return data
+
+
+def _needs_rotation(var_info: VarInfoFromNetCDF4, variable: str) -> bool:
+    """Check if variable must be rotated before resampling.
+
+    pyresample's EWA assumes swath input which implies the x projection
+    dimension must be the fastest varying dimension.
+
+    So if the lon comes before lat in the variables dimensions you must rotate
+    the grid before resampling.
+
+    """
+    needs_rotation = False
+    var_dims = var_info.get_variable(variable).dimensions
+    xloc = next(
+        (
+            index
+            for index, dimension in enumerate(var_dims)
+            if _is_column_dim(dimension, var_info)
+        ),
+        None,
+    )
+    yloc = next(
+        (
+            index
+            for index, dimension in enumerate(var_dims)
+            if _is_row_dim(dimension, var_info)
+        ),
+        None,
+    )
+    if yloc > xloc:
+        needs_rotation = True
+
+    return needs_rotation
