+++ harmony_regridding_service/utilities.py
@@ -1,26 +1,33 @@
 """Utility functions for use within the Harmony Regridding Service.
 
-These include MIME type determination and basic components of message
-validation.
+This module contains lower-level functionality that can be abstracted into
+functions. Primarily this improves readability of the source code, and
+allows finer-grained unit testing of each smaller part of functionality.
 
 """
 
+from logging import getLogger
 from mimetypes import guess_type as guess_mime_type
 from os.path import splitext
+from pathlib import PurePath
 
-from harmony_service_lib.message import Message
-from harmony_service_lib.message_utility import rgetattr
-from pyproj import CRS
-from pyproj.exceptions import CRSError
+import numpy as np
+from netCDF4 import (
+    Dataset,
+    Group,
+    Variable,
+)
+from varinfo import VarInfoFromNetCDF4
 
-from harmony_regridding_service.exceptions import InvalidTargetCRS
+from harmony_regridding_service.exceptions import SourceDataError
+
+logger = getLogger(__name__)
 
 KNOWN_MIME_TYPES = {
     '.nc4': 'application/x-netcdf4',
     '.h5': 'application/x-hdf5',
     '.hdf5': 'application/x-hdf5',
 }
-VALID_INTERPOLATION_METHODS = ('Elliptical Weighted Averaging',)
 
 
 def get_file_mime_type(file_name: str) -> str | None:
@@ -41,47 +48,116 @@ def get_file_mime_type(file_name: str) -> str | None:
     return mime_type[0]
 
 
-def has_valid_crs(message: Message) -> bool:
-    """Validate CRS.
+def _walk_groups(node: Dataset | Group) -> Group:
+    """Traverse a netcdf file yielding each group."""
+    yield node.groups.values()
+    for value in node.groups.values():
+        yield from _walk_groups(value)
 
-    Check the target Coordinate Reference System (CRS) in the Harmony
-    message is compatible with the CRS types supported by the regridder. In
-    the MVP, only a geographic CRS is supported, so `Message.format.crs`
-    should either be undefined or specify a geographic CRS.
+
+def _transfer_metadata(source_ds: Dataset, target_ds: Dataset) -> None:
+    """Transfer over global and group metadata to target file."""
+    global_metadata = {}
+    for attr in source_ds.ncattrs():
+        global_metadata[attr] = source_ds.getncattr(attr)
+
+    target_ds.setncatts(global_metadata)
+
+    for groups in _walk_groups(source_ds):
+        for group in groups:
+            group_metadata = {}
+            for attr in group.ncattrs():
+                group_metadata[attr] = group.getncattr(attr)
+            t_group = target_ds.createGroup(group.path)
+            t_group.setncatts(group_metadata)
+
+
+def _integer_like(test_type: np.dtype) -> bool:
+    """Return True if the datatype is integer like."""
+    return np.issubdtype(np.dtype(test_type), np.integer)
+
+
+def _copy_var_with_attrs(
+    source_ds: Dataset, target_ds: Dataset, variable_name: str
+) -> tuple[Variable, Variable]:
+    """Copy a source variable and metadata to target.
+
+    Copy both the variable and metadata from a souce variable into a target,
+    return both source and target variables.
+    """
+    s_var, t_var = _copy_var_without_metadata(source_ds, target_ds, variable_name)
+
+    for att in s_var.ncattrs():
+        if att != '_FillValue':
+            t_var.setncattr(att, s_var.getncattr(att))
+
+    return (s_var, t_var)
+
+
+def _copy_var_without_metadata(
+    source_ds: Dataset, target_ds: Dataset, variable_name: str
+) -> tuple[Variable, Variable]:
+    """Clones a single variable and returns both source and target variables.
+
+    This function uses the netCDF4 createGroup('/[optionalgroup/andsubgroup]')
+    call This will return an existing group, or create one that does not
+    already exists. So this is not clobbering the source data.
 
     """
-    target_crs = rgetattr(message, 'format.crs')
-    return target_crs is None or _is_geographic_crs(target_crs)
+    var = PurePath(variable_name)
+    s_var = _get_variable(source_ds, variable_name)
+
+    # Create target variable
+    t_group = target_ds.createGroup(var.parent)
+    fill_value = getattr(s_var, '_FillValue', None)
+    t_var = t_group.createVariable(
+        var.name, s_var.dtype, s_var.dimensions, fill_value=fill_value
+    )
+    s_var.set_auto_maskandscale(False)
+    t_var.set_auto_maskandscale(False)
+
+    return (s_var, t_var)
 
 
-def _is_geographic_crs(crs_string: str) -> bool:
-    """Infer if CRS is geographic.
+def _clone_variables(
+    source_ds: Dataset, target_ds: Dataset, variables: set[str]
+) -> set[str]:
+    """Clone variables from source to target.
 
-    Use pyproj to ascertain if the supplied Coordinate Reference System
-    (CRS) is geographic.
+    Copy variables and their attributes directly from the source Dataset to the
+    target Dataset.
     """
-    try:
-        crs = CRS(crs_string)
-        is_geographic = crs.is_geographic
-    except CRSError as exception:
-        raise InvalidTargetCRS(crs_string) from exception
+    for variable_name in variables:
+        (s_var, t_var) = _copy_var_with_attrs(source_ds, target_ds, variable_name)
+        try:
+            t_var[:] = s_var[:]
+        except IndexError as vlen_error:
+            # Handle snowflake metadata with vlen string.
+            if s_var.dtype == str and s_var.shape == ():
+                t_var[0] = s_var[0]
+            else:
+                logger.error('Unable to clone variable {s_var}')
+                raise SourceDataError('Unhandled variable clone') from vlen_error
 
-    return is_geographic
+    return variables
 
 
-def has_valid_interpolation(message: Message) -> bool:
-    """Ensure valid interpolation.
+def _get_variable(dataset: Dataset, variable_name: str) -> Variable:
+    """Return a variable from a fully qualified variable name.
 
-    Check the interpolation method in the input Harmony message is
-    compatible with the methods supported by the regridder. In the MVP,
-    only the EWA algorithm is used, so either the interpolation should be
-    unspecified in the message, or it should be the string
-    "Elliptical Weighted Averaging".
-
+    This will return an existing or create a new variable.
     """
-    interpolation_method = rgetattr(message, 'format.interpolation')
+    var = PurePath(variable_name)
+    group = dataset.createGroup(var.parent)
+    return group[var.name]
 
-    return (
-        interpolation_method is None
-        or interpolation_method in VALID_INTERPOLATION_METHODS
+
+def _get_bounds_var(var_info: VarInfoFromNetCDF4, dim_name: str) -> str:
+    return next(
+        (
+            var_info.get_variable(f'{dim_name}_{ext}').name
+            for ext in ['bnds', 'bounds']
+            if var_info.get_variable(f'{dim_name}_{ext}') is not None
+        ),
+        None,
     )
